chore: merge test-helpers.js
